refactor : 카운트 타운 화면 대신 스크립트화
